Refactored, made ProjectContainer+Project classes
Refactored a lot of files trying to make things more readable
    - makeAll.js
    - makeSidebarElems
    - addNew
    - addNewFunctions
    - sidebarTabCreation
    - taskItemHover
    - taskItemHoverFunctions
    - getDate (this one is a lot better...)

Created ProjectContainer and Project classes.
    - Make a project, push it to ProjectContainer
    - projectContainer.display() creates the html
    showing incomplete vs complete
    - allTaskManager contains all tasks for the overview page

Need to:
find a way to update projects on finishing or deleting a task
Create a way to edit the tasks after creation.

NOTE: taskObj was changed to allTaskManager.
class for allTaskManager was changed to TaskManager